Moved books logic to controller
