+++ routes/books.js
@@ -3,182 +3,25 @@ const router = express.Router();
 const path = require("path");
 const db = require("../models");
 const pdf = require("html-pdf");
+const Book = require("../controllers/booksController");
 const isAuth = require("../middleware/authMiddleware").isAuth;
-const isNotAuth = require("../middleware/authMiddleware").isNotAuth;
-const mailer = require("../email/mailer");
-const getTemplate = require("../helpers/email-template").getTemplate;
 
-router.get("/add", isAuth, async (req, res, next) => {
-  try {
-    res.render("books/add", { loggedin: req.isAuthenticated() ? true : false });
-    res.status(200);
-  } catch (err) {
-    next(err);
-  }
-});
+router.get("/add", isAuth, Book.add);
 
-router.post("/add", isAuth, async (req, res, next) => {
-  try {
-    const user = await req.user;
-    const { title, author, publication_date, abstract, cover } = req.body;
-    const userId = await user.id;
-    const book = await db.Book.create({
-      title,
-      author,
-      publication_date,
-      abstract,
-      cover,
-      userId,
-    });
-    const htmlToSend = getTemplate("newBook", {
-      username: user.username,
-      book: book,
-    });
-    mailer.sendMail(htmlToSend, user);
-    res.redirect("/books");
-  } catch (err) {
-    next(err);
-  }
-});
+router.post("/add", isAuth, Book.create);
 
-router.get("/generate-report", isAuth, async (req, res, next) => {
-  try {
-    const user = await req.user;
-    const books = await db.Book.findAll({ where: { userId: user.id } });
-    const pdfOptions = {
-      height: "11.25in",
-      width: "8.5in",
-      header: {
-        height: "20mm",
-      },
-      footer: {
-        height: "20mm",
-      },
-    };
-    res.render(
-      "pdf/report-template",
-      {
-        username: user.username,
-        books: books,
-        noHeader: true,
-        date: new Date().toDateString(),
-      },
-      (err, html) => {
-        if (err) {
-          next(err);
-        } else {
-          pdf
-            .create(html, pdfOptions)
-            .toFile("views/pdf/report.pdf", (err, data) => {
-              if (err) {
-                next(err);
-              } else {
-                res.send("Report created");
-                res.status(200);
-              }
-            });
-        }
-      }
-    );
-  } catch (err) {
-    next(err);
-  }
-});
+router.get("/generate-report", isAuth, Book.getReport);
 
-router.get("/edit/:id", isAuth, async (req, res, next) => {
-  try {
-    let id = req.params.id;
-    const book = await db.Book.findByPk(id);
-    res.render("books/edit", { book });
-    res.status(200);
-  } catch (err) {
-    next(err);
-  }
-});
+router.get("/edit/:id", isAuth, Book.edit);
 
-router.post("/edit/:id", isAuth, async (req, res, next) => {
-  try {
-    let id = req.params.id;
-    const { title, author, publication_date, abstract, cover } = req.body;
-    await db.Book.update(
-      {
-        title,
-        author,
-        publication_date,
-        abstract,
-        cover,
-      },
-      { where: { id } }
-    );
-    res.redirect("/books");
-    res.status(200);
-  } catch (err) {
-    next(err);
-  }
-});
+router.post("/edit/:id", isAuth, Book.update);
 
-router.get("/destroy/confirm/:id", isAuth, async (req, res, next) => {
-  try {
-    let id = req.params.id;
-    const book = await db.Book.findByPk(id);
-    res.render("books/delete", {
-      title: book.title,
-      book,
-    });
-  } catch (err) {
-    next(err);
-  }
-});
+router.get("/destroy/confirm/:id", isAuth, Book.preDestroy);
 
-router.post("/destroy/confirm/:id", isAuth, async (req, res, next) => {
-  try {
-    let id = req.params.id;
-    const user = await req.user;
-    const book = await db.Book.findByPk(id);
-    await book.destroy();
-    const htmlToSend = getTemplate("deletedBook", {
-      username: user.username,
-      book: book,
-    });
-    mailer.sendMail(htmlToSend, user);
-    res.status(200);
-    res.redirect("/");
-  } catch (err) {
-    next(err);
-  }
-});
+router.post("/destroy/confirm/:id", isAuth, Book.destroy);
 
-router.get("/:id", isAuth, async (req, res, next) => {
-  try {
-    const id = req.params.id;
-    const user = await req.user;
-    const userId = user.id;
-    const book = await db.Book.findByPk(id);
-    if (book !== null && book.userId !== userId) {
-      req.flash("info", "You can interactuare just with your Books");
-      res.redirect("/books");
-    }
-    res.status(200);
-    res.render("books/view", { book: book });
-  } catch (err) {
-    next(err);
-  }
-});
+router.get("/:id", isAuth, Book.read);
 
-router.get("/", isAuth, async (req, res, next) => {
-  try {
-    const user = await req.user;
-    const userId = await user.id;
-    const books = await db.Book.findAll({ where: { userId: userId } });
-    res.status(200);
-    res.render("books/index", {
-      title: "Books",
-      books: books,
-      loggedin: req.isAuthenticated() ? true : false,
-    });
-  } catch (err) {
-    next(err);
-  }
-});
+router.get("/", isAuth, Book.all);
 
 module.exports = router;
